% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sentiment_score.R
\name{aggregate_sentiment_score}
\alias{aggregate_sentiment_score}
\title{Returns an aggregated compound score representing sentiment:
-1 (most extreme negative) and +1 (most extreme positive). The compound score
is a normalized score calculated by summing the valence scores of each word
in the lexicon.}
\usage{
aggregate_sentiment_score(df, col)
}
\arguments{
\item{df, }{data frame containing the reviews}

\item{col, }{name of the column that contains the reviews}
}
\value{
float value between -1 to 1
}
\description{
Returns an aggregated compound score representing sentiment:
-1 (most extreme negative) and +1 (most extreme positive). The compound score
is a normalized score calculated by summing the valence scores of each word
in the lexicon.
}
\examples{
score1 = aggregate_sentiment_score(df, "text")
score2 = aggregate_sentiment_score(df, "review")
}
