% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_compound_score}
\alias{get_compound_score}
\title{get_compound_score
Calculates a compound sentiment score from text}
\usage{
get_compound_score(text)
}
\arguments{
\item{text, }{Input text data as string}
}
\value{
A numeric sentiment score of the text,
ranges from -1 for negative
0 for neutral and 1 for positive
}
\description{
get_compound_score
Calculates a compound sentiment score from text
}
\examples{
get_compound_score("This is great")
get_compound_score("This is a neutral sentence")
}
