% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_sentiment}
\alias{get_sentiment}
\title{get_sentiment function
Evaluates the type of sentiment from a given text}
\usage{
get_sentiment(text)
}
\arguments{
\item{text, }{Input text data as string}
}
\value{
The type of sentiment (positive, negative or neutral)
as a character
}
\description{
get_sentiment function
Evaluates the type of sentiment from a given text
}
\examples{
get_sentiment("This is great")
get_sentiment("This is a neutral sentence")
}
