% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_sentiment_and_score}
\alias{get_sentiment_and_score}
\title{get_sentiment_and_score
Adds the sentiment and score from text to data frame}
\usage{
get_sentiment_and_score(df, col)
}
\arguments{
\item{df, }{dataframe}

\item{col, }{name of the text column on which sentiment analysis is done}
}
\value{
A dataframe with numeric sentiment and its sentiment type added,
to the dataframe
}
\description{
get_sentiment_and_score
Adds the sentiment and score from text to data frame
}
\examples{
df =  data.frame(text = c("I love cats","I love dogs"))
get_sentiment_and_score(df, "text")
}
