% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_score_plot.R
\name{sentiment_score_plot}
\alias{sentiment_score_plot}
\title{sentiment_score_plot}
\usage{
sentiment_score_plot(df, col)
}
\arguments{
\item{df, }{data frame containing the reviews}

\item{col, }{name of the column that contains the reviews}
}
\value{
plot object
}
\description{
Creates a binned histogram showing count of reviews
ranging from negative to positive sentiment score.
}
\details{
Provides the most frequent sentiment score and the
corresponding percentage
}
\examples{
plot1 <- sentiment_score_plot(df, "column_name")
plot2 <- sentiment_score_plot(df, "reviews")

}
